Added ranged attack feature for enemy units

+++ src/mercenary.c
@@ -230,7 +230,7 @@ void mercenary_reassignment_selection(int listing_selected) {
   to a trade ship or the player's ship
 */
 void mercenary_reassign(void *ch) {
-  long change = (long) ch;
+  size_t change = (size_t) ch;
   if ((e_player.total_mercenaries && change > 0) ||
        change < 0) {
     int is_selected = -1;
